goldybot/requirements.txt
discord
disnake
pyttsx3
discord.py[voice]
Cybernator
rapidfuzz
fuzzywuzzy
pillow
random
sql3